check if DNS is allowed if needed
If one of the ALLOW or BLOCK rules specify a glob that will require
a DNS request to check access (e.g. `*.1e100.net') then we ensure
first that this DNS request itself will be allowed.

This fixes a long standing bug where an infinite loop was triggered
trying to resolve a host but DNS connection was not allowed. Now we
prevent execution in such situations, reporting:

  No DNS allowed while some glob rule need one, aborting