fix a segfault occurring with SELinux enabled
For audit purpose when SELinux is enabled `connect' gets called while
the `coc_init' function has not yet been called.

In that case `real_connect' was NULL, leading to the segfault.

We work around the issue by ensuring `real_connect' is initialized when
`connect' is called before connect-or-cut is initialized.

The fix assumes that the machinery involved does not require mutex protection.