initial version of IPv6 introduction
Along with this change we are simplifying data structures and now rely on a
single linked list.

This will enable interleaving ALLOW and BLOCK rules, and the helper shell
script will evolve to take this into account.

This commit breaks some corner cases where we were checking for
inconsistencies. These checks have been temporarily disabled.

I'm committing this into a separate branch because this has not been tested
yet; it compiles only.