WIN32: handle 32-bit injection from 64-bit
Our injection mechanism does not work if we use 64-bit coc.exe to run a 32-bit binary, or if we use coc32.exe with a 64-bit binary. These use cases are broken because:
 - we assume LoadLibrary address to be the same in our coc process and in the started one
 - we use WriteProcessMemory and CreateRemoteThread for a given kernel32.dll, but in fact we should be using their other bitness variant.

Here we perform a naïve approach: in such "mixed" circumstances we call the process to start with the correct coc injector binary:
 - coc.exe for 64-bit binaries
 - coc32.exe for 32-bit binaries

This is a first round. It needs to be refactored and used in connect-or-cut itself.

At the moment it's not very clever because we start the process and then check its bitness. If it's not ok we kill what we just started and use coc.exe (or coc32.exe) to perform injection that will work.

I've found various writeups  to mix 64-bit and 32-bit like the following:
  http://blog.rewolf.pl/blog/?p=102
but in our case since we are hijacking CreateProcess we can use coc as a trampoline again if need be to avoid this extra-complexity.