WIN32: refactor LoadCoCLibrary to handle 32/64 mixup
The function now uses a struct holding all CreateProcess arguments and a pointer to a CreateProcess compatible function.

It's used both in coc and in the CreateProcess hook.

Tested scenarios:
 1. running a 32-bit process with coc.exe is OK(process is relaunched with coc32.exe)
 2. running a 32-bit process from an already injected 64-bit binary (cmd.exe) is KO

This latter scenario is the one that triggered this refactoring: it should be possible to run a 32-bit process from the hooked CreateProcess and have the 32-bit process reran with coc32.exe like we're doing in bootstrap.