WIN32: Repair CreateProcess when injected
We were zeroing memory of STARTUPINFO and
this caused the invalid argument (87) error.

Also make sure to start the process in a
non-suspended way when re-running the injector
to change the bitness.

Last, add a missing NULL for environment that
would have blow in release mode for sure.

With this commit we can successfully:
 - bootstrap 32-bit process from 64-bit coc.exe
 - run 32-bit process from 64-bit injected process

We still need to test the opposite: from 32 to 64.

To be investigated later: testing under
cmd.exe shows weird display behaviour.