WIN32: revamp injection and CreateProcess hook
We now check whether image is same bitness
as the current running process. If they differ then
we run new process with coc.exe (or coc32.exe).

We do this before trying to run the process to
avoid running then terminating once started.
This relies on Imagehlp library to check for
the PE about to be run.

On top of that we now also look if the process
is a console one, and in that case we wait for its
completion.

The weird display and command prompt
intermixed disappears thanks to this last
change.